requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
numpy==1.24.3
pinecone-client==3.0.0
tensorflow==2.15.0
scikit-image==0.22.0
Pillow==10.1.0